+++ src/main/java/power/keepeersofthestones/procedures/LibraryBuildProcedure.java
@@ -43,7 +43,7 @@ public class LibraryBuildProcedure {
 		if (world instanceof ServerWorld) {
 			Template template = ((ServerWorld) world).getStructureTemplateManager().getTemplateDefaulted(new ResourceLocation("power", "library"));
 			if (template != null) {
-				template.func_237144_a_((ServerWorld) world, new BlockPos((int) new Object() {
+				template.func_237144_a_((ServerWorld) world, new BlockPos(new Object() {
 					double convert(String s) {
 						try {
 							return Double.parseDouble(s.trim());
@@ -59,7 +59,7 @@ public class LibraryBuildProcedure {
 						}
 						return "";
 					}
-				}.getText()), (int) new Object() {
+				}.getText()), new Object() {
 					double convert(String s) {
 						try {
 							return Double.parseDouble(s.trim());
@@ -75,7 +75,7 @@ public class LibraryBuildProcedure {
 						}
 						return "";
 					}
-				}.getText()), (int) new Object() {
+				}.getText()), new Object() {
 					double convert(String s) {
 						try {
 							return Double.parseDouble(s.trim());
