+++ src/main/java/power/keepeersofthestones/procedures/OpenBookProcedure.java
@@ -120,7 +120,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -139,7 +139,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -158,7 +158,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -177,7 +177,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -196,7 +196,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -215,7 +215,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -234,7 +234,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -253,7 +253,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -272,7 +272,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -291,7 +291,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -310,7 +310,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -329,7 +329,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -348,7 +348,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -367,7 +367,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -386,7 +386,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -405,7 +405,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -424,7 +424,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -443,7 +443,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -462,7 +462,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -481,7 +481,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -500,7 +500,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -519,7 +519,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -538,7 +538,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -557,7 +557,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -576,7 +576,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -595,7 +595,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -614,7 +614,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -633,7 +633,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -652,7 +652,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -671,7 +671,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
@@ -690,7 +690,7 @@ public class OpenBookProcedure {
 			{
 				Entity _ent = entity;
 				if (_ent instanceof ServerPlayerEntity) {
-					BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+					BlockPos _bpos = new BlockPos(x, y, z);
 					NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
 						@Override
 						public ITextComponent getDisplayName() {
